refactor: update import paths for typing modules to use the new naming convention
