+++ lionagi/core/models/schema_model.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from ..typing.pydantic_ import ConfigDict
+from ..typing._pydantic import ConfigDict
 from .base import BaseAutoModel
 
 __all__ = ("SchemaModel",)
